format code

+++ vllm/model_executor/layers/quantization/compressed_tensors/compressed_tensors.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from contextlib import suppress
-from typing import Callable, Any, Literal, Optional, cast
+from typing import Any, Callable, Literal, Optional, cast
 
 import torch
 from compressed_tensors.config import (CompressionFormat,
@@ -620,16 +620,18 @@ class CompressedTensorsLinearMethod(LinearMethodBase):
             raise ValueError("A scheme must be defined for each layer")
         return scheme.apply_weights(layer, x, bias=bias)
 
-    def dequant_fp8_weight(self, layer: "CompressedTensorsW8A8Fp8") -> torch.Tensor:
+    def dequant_fp8_weight(self,
+                           layer: "CompressedTensorsW8A8Fp8") -> torch.Tensor:
         if layer.scheme.strategy == QuantizationStrategy.CHANNEL.value:
-            dequant_weight = layer.weight.to(layer.weight_scale.dtype) * layer.weight_scale.squeeze()
+            dequant_weight = layer.weight.to(
+                layer.weight_scale.dtype) * layer.weight_scale.squeeze()
             return dequant_weight.to(torch.bfloat16).t()
         else:
-            raise NotImplementedError("Implemented per-channel dequantization only")
+            raise NotImplementedError(
+                "Implemented per-channel dequantization only")
 
     def get_dequant_weights_func(
-        self,
-    ) -> Optional[Callable[[torch.nn.Module], torch.Tensor]]:
+        self, ) -> Optional[Callable[[torch.nn.Module], torch.Tensor]]:
         return self.dequant_fp8_weight
